refactor(lib): split github.ts into modular folder, add vitest
Decompose the 588-line monolithic github.ts into a focused folder:
- client.ts — shared headers, githubJson(), searchMergedPrCount() helpers
- schemas.ts — zod schemas
- star-history.ts — pure buildStarHistory() function
- contributions.ts / stars.ts / ecosystem.ts — feature-grouped API fns
- index.ts — barrel preserving all existing @/lib/github imports

Add vitest with two unit test suites in tests/lib/:
- star-history.test.ts (downsampling math, edge cases)
- http.test.ts (HttpError, isRetryableHttpError)
