+++ vitest.config.ts
@@ -1,0 +1,14 @@
+import { resolve } from "node:path";
+import { defineConfig } from "vitest/config";
+
+export default defineConfig({
+  resolve: {
+    alias: {
+      "@": resolve(__dirname, "./src"),
+    },
+  },
+  test: {
+    environment: "node",
+    include: ["tests/**/*.test.{ts,tsx}"],
+  },
+});
